create ImageStorageManager
moved methods to store and display images to separate class
