+++ src/QRCheckIn/app/src/main/java/com/example/qrcheckin/EventDatabaseManager.java
@@ -6,6 +6,9 @@ import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 
+/**
+ * Controls storing and retrieving data from the events firebase collection
+ */
 public class EventDatabaseManager {
     private final FirebaseFirestore db;
     private final CollectionReference eventCollectionRef;
@@ -34,7 +37,12 @@ public class EventDatabaseManager {
     public CollectionReference getEventCollectionRef(){
         return this.eventCollectionRef;
     }
-    
+
+    /**
+     * Returns a DocumentReference for an Event document
+     * @param documentId doc ID String of the event in firebase
+     * @return DocumentReference of the Event
+     */
     public DocumentReference getEventDocRef(String documentId){
         return eventCollectionRef.document(documentId);
     }
